+++ lib/broadway_dashboard/new_telemetry.ex
@@ -1,4 +1,4 @@
-defmodule BroadwayDashboard.NewTelemetry do
+defmodule BroadwayDashboard.Telemetry do
   @moduledoc false
 
   # Measurements of pipelines are based on telemetry events.
@@ -8,7 +8,7 @@ defmodule BroadwayDashboard.NewTelemetry do
   # If this time is shorter, it means that the stage is
   # doing more work.
 
-  alias BroadwayDashboard.NewCounters
+  alias BroadwayDashboard.Counters
 
   def attach(parent, pipeline, counters) do
     events = [
@@ -44,14 +44,14 @@ defmodule BroadwayDashboard.NewTelemetry do
       when stage_layer in [:processor, :batcher, :consumer] do
     case stage_layer do
       :processor ->
-        :ok = NewCounters.put_processor_start(counters, metadata.index, measurements.time)
+        :ok = Counters.put_processor_start(counters, metadata.index, measurements.time)
 
       :batcher ->
-        :ok = NewCounters.put_batcher_start(counters, metadata.batcher_key, measurements.time)
+        :ok = Counters.put_batcher_start(counters, metadata.batcher_key, measurements.time)
 
       :consumer ->
         :ok =
-          NewCounters.put_batch_processor_start(
+          Counters.put_batch_processor_start(
             counters,
             metadata.batch_info.batcher,
             metadata.index,
@@ -69,46 +69,46 @@ defmodule BroadwayDashboard.NewTelemetry do
       when stage_layer in [:processor, :batcher, :consumer] do
     case stage_layer do
       :processor ->
-        {:ok, start_time} = NewCounters.get_processor_start(counters, metadata.index)
-        {:ok, last_end_time} = NewCounters.get_processor_end(counters, metadata.index)
+        {:ok, start_time} = Counters.get_processor_start(counters, metadata.index)
+        {:ok, last_end_time} = Counters.get_processor_end(counters, metadata.index)
 
         factor = calc_factor(start_time, last_end_time, measurements.duration)
 
-        :ok = NewCounters.put_processor_end(counters, metadata.index, measurements.time)
-        :ok = NewCounters.put_processor_processing_factor(counters, metadata.index, factor)
+        :ok = Counters.put_processor_end(counters, metadata.index, measurements.time)
+        :ok = Counters.put_processor_processing_factor(counters, metadata.index, factor)
 
         # Here we measure only because it can occur a failure or
         # we don't have batchers and we "Ack" in the processor.
         :ok =
-          NewCounters.incr(
+          Counters.incr(
             counters,
             length(metadata.successful_messages_to_ack),
             length(metadata.failed_messages)
           )
 
       :batcher ->
-        {:ok, start_time} = NewCounters.get_batcher_start(counters, metadata.batcher_key)
-        {:ok, last_end_time} = NewCounters.get_batcher_end(counters, metadata.batcher_key)
+        {:ok, start_time} = Counters.get_batcher_start(counters, metadata.batcher_key)
+        {:ok, last_end_time} = Counters.get_batcher_end(counters, metadata.batcher_key)
 
         factor = calc_factor(start_time, last_end_time, measurements.duration)
 
-        :ok = NewCounters.put_batcher_end(counters, metadata.batcher_key, measurements.time)
-        :ok = NewCounters.put_batcher_processing_factor(counters, metadata.batcher_key, factor)
+        :ok = Counters.put_batcher_end(counters, metadata.batcher_key, measurements.time)
+        :ok = Counters.put_batcher_processing_factor(counters, metadata.batcher_key, factor)
 
       :consumer ->
         key = metadata.batch_info.batcher
         index = metadata.index
 
-        {:ok, start_time} = NewCounters.get_batch_processor_start(counters, key, index)
-        {:ok, last_end_time} = NewCounters.get_batch_processor_end(counters, key, index)
+        {:ok, start_time} = Counters.get_batch_processor_start(counters, key, index)
+        {:ok, last_end_time} = Counters.get_batch_processor_end(counters, key, index)
 
         factor = calc_factor(start_time, last_end_time, measurements.duration)
 
-        :ok = NewCounters.put_batch_processor_end(counters, key, index, measurements.time)
-        :ok = NewCounters.put_batch_processor_processing_factor(counters, key, index, factor)
+        :ok = Counters.put_batch_processor_end(counters, key, index, measurements.time)
+        :ok = Counters.put_batch_processor_processing_factor(counters, key, index, factor)
 
         :ok =
-          NewCounters.incr(
+          Counters.incr(
             counters,
             length(metadata.successful_messages),
             length(metadata.failed_messages)
@@ -122,7 +122,7 @@ defmodule BroadwayDashboard.NewTelemetry do
         %{topology_name: pipeline} = _metadata,
         {pipeline, counters}
       ) do
-    :ok = NewCounters.incr(counters, 0, 1)
+    :ok = Counters.incr(counters, 0, 1)
   end
 
   # Ignore events from other pipelines
